project commit

+++ src/main/java/org/example/appproject_be/controller/HardwareController.java
@@ -1,5 +1,7 @@
 package org.example.appproject_be.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,13 +25,23 @@ public class HardwareController {
     private final HardwareService hardwareService;
 
     @PostMapping("/assets/hardware")
-    public HardwareDto saveAsset(@Valid @RequestBody HardwareDto hardware) {
+    public HardwareDto saveAsset(@Valid @RequestBody HardwareDto hardware, HttpServletRequest request) {
+        // 세션 가져오기
+        HttpSession session = request.getSession(false); // false로 설정하면 세션이 없으면 null을 반환
+        // 세션이 null이면 에러 반환
+        if (session == null) {
+            throw new IllegalStateException("해당 작업을 수행하기 위한 권한이 없습니다. 관리자 계정으로 로그인해주세요.");
+        }
         return hardwareService.saveHardware(hardware);
     }
 
     @DeleteMapping("/assets/hardware")
-    public ResponseEntity<String> deleteAsset(@RequestParam("idx") Long idx) {
-
+    public ResponseEntity<String> deleteAsset(@RequestParam("idx") Long idx, HttpServletRequest request) {
+        HttpSession session = request.getSession(false); // false로 설정하면 세션이 없으면 null을 반환
+        // 세션이 null이면 에러 반환
+        if (session == null) {
+            throw new IllegalStateException("해당 작업을 수행하기 위한 권한이 없습니다. 관리자 계정으로 로그인해주세요.");
+        }
         hardwareService.deleteHardware(idx);
         return new ResponseEntity<>("해당 자산을 삭제하였습니다.", HttpStatus.OK);
     }
@@ -49,7 +61,12 @@ public class HardwareController {
     }
 
     @PutMapping("/assets/hardware/update")
-    public HardwareDto updateAsset(@RequestParam("hwidx") Long hwidx, @RequestParam("assetidx") Long assetidx,  @RequestBody HardwareDto hardwareDto) {
+    public HardwareDto updateAsset(@RequestParam("hwidx") Long hwidx, @RequestParam("assetidx") Long assetidx,  @RequestBody HardwareDto hardwareDto, HttpServletRequest request) {
+        HttpSession session = request.getSession(false); // false로 설정하면 세션이 없으면 null을 반환
+        // 세션이 null이면 에러 반환
+        if (session == null) {
+            throw new IllegalStateException("해당 작업을 수행하기 위한 권한이 없습니다. 관리자 계정으로 로그인해주세요.");
+        }
         hardwareDto.setHwidx(hwidx);
         hardwareDto.setAssetidx(assetidx);
         return hardwareService.updateHardware(hardwareDto);
