+++ src/main/java/org/example/appproject_be/controller/SoftwareController.java
@@ -1,5 +1,7 @@
 package org.example.appproject_be.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,12 +25,22 @@ public class SoftwareController {
     private final SoftwareService softwareService;
     //localhost:8080/assets/software
     @PostMapping("/assets/software")
-    public SoftwareDto saveAsset(@Valid @RequestBody SoftwareDto software){
+    public SoftwareDto saveAsset(@Valid @RequestBody SoftwareDto software, HttpServletRequest request){
+        HttpSession session = request.getSession(false); // false로 설정하면 세션이 없으면 null을 반환
+        // 세션이 null이면 에러 반환
+        if (session == null) {
+            throw new IllegalStateException("해당 작업을 수행하기 위한 권한이 없습니다. 관리자 계정으로 로그인해주세요.");
+        }
         return softwareService.saveSoftware(software);
     }
     //localhost:8080/assets/software?idx=idx2123
     @DeleteMapping("/assets/software")
-    public ResponseEntity<String> deleteAsset (@RequestParam("idx") Long idx){
+    public ResponseEntity<String> deleteAsset (@RequestParam("idx") Long idx, HttpServletRequest request){
+        HttpSession session = request.getSession(false); // false로 설정하면 세션이 없으면 null을 반환
+        // 세션이 null이면 에러 반환
+        if (session == null) {
+            throw new IllegalStateException("해당 작업을 수행하기 위한 권한이 없습니다. 관리자 계정으로 로그인해주세요.");
+        }
 
         softwareService.deleteSoftware(idx);
         return new ResponseEntity<>("해당 자산을 삭제하였습니다.", HttpStatus.OK);
@@ -43,7 +55,12 @@ public class SoftwareController {
         return softwareService.getSoftware(idx);
     }
     @PutMapping("/assets/software/update")
-    public SoftwareDto updateAsset(@RequestParam("swidx") Long swidx, @RequestParam("assetidx") Long assetidx,  @RequestBody SoftwareDto softwareDto) {
+    public SoftwareDto updateAsset(@RequestParam("swidx") Long swidx, @RequestParam("assetidx") Long assetidx,  @RequestBody SoftwareDto softwareDto, HttpServletRequest request) {
+        HttpSession session = request.getSession(false); // false로 설정하면 세션이 없으면 null을 반환
+        // 세션이 null이면 에러 반환
+        if (session == null) {
+            throw new IllegalStateException("해당 작업을 수행하기 위한 권한이 없습니다. 관리자 계정으로 로그인해주세요.");
+        }
         softwareDto.setSwidx(swidx);
         softwareDto.setAssetidx(assetidx);
         return softwareService.updateSoftware(softwareDto);
